+++ packages/dsw-tdk/CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [3.20.0]
+
+### Added
+
+- Added `unpackage` command
+
+### Changed
+
+- Updated according to API changes for Template Editor
+
 ## [3.19.2]
 
 ### Fixed
@@ -263,3 +273,4 @@ Initial DSW Template Development Kit (versioned as part of the [DSW platform](ht
 [3.19.0]: /../../tree/v3.19.0
 [3.19.1]: /../../tree/v3.19.1
 [3.19.2]: /../../tree/v3.19.2
+[3.20.0]: /../../tree/v3.20.0